Create Comparing an array against another array.
This is an example of splitting a string of words into an array of words, and comparing that against another array of words, to see if any of the words in one array appear in any of the words in the other array. 
It can be used for a wide range of things.