Create Comparing an array against another array2.
This is an example of splitting a string of words into an array of words, and comparing that against another array of words, to see if any of the words in one array appear in any of the words in the other array. 
This script has a bunch of other functions involved relating to the formatting of the text that arent associated with comparing one array against another. Clean the code to be relevant to your use before deploying. 
It can be used for a wide range of things.